+++ packages/matter.js/src/protocol/ExchangeManager.ts
@@ -32,6 +32,8 @@ import { SecureChannelMessenger } from "./securechannel/SecureChannelMessenger.j
 
 const logger = Logger.get("ExchangeManager");
 
+const MAXIMUM_CONCURRENT_EXCHANGES_PER_SESSION = 5;
+
 export class ChannelNotConnectedError extends MatterError {}
 
 export class MessageChannel<ContextT extends SessionContext> implements Channel<Message> {
@@ -284,9 +286,6 @@ export class ExchangeManager<ContextT extends SessionContext> {
                     );
                 }
             }
-
-            // TODO A node SHOULD limit itself to a maximum of 5 concurrent exchanges over a unicast session. This is
-            //  to prevent a node from exhausting the message counter window of the peer node.
         }
     }
 
@@ -358,6 +357,24 @@ export class ExchangeManager<ContextT extends SessionContext> {
     #addExchange(exchangeIndex: number, exchange: MessageExchange<ContextT>) {
         exchange.closed.on(() => this.deleteExchange(exchangeIndex));
         this.exchanges.set(exchangeIndex, exchange);
+
+        // A node SHOULD limit itself to a maximum of 5 concurrent exchanges over a unicast session. This is
+        // to prevent a node from exhausting the message counter window of the peer node.
+        // TODO Make sure Group sessions are handled differently
+        this.#cleanupSessionExchanges(exchange.session.id);
+    }
+
+    #cleanupSessionExchanges(sessionId: number) {
+        const sessionExchanges = Array.from(this.exchanges.values()).filter(
+            exchange => exchange.session.id === sessionId && !exchange.isClosing,
+        );
+        if (sessionExchanges.length <= MAXIMUM_CONCURRENT_EXCHANGES_PER_SESSION) {
+            return;
+        }
+        // let's use the first entry in the Map as the oldest exchange and close it
+        const exchangeToClose = sessionExchanges[0];
+        logger.debug(`Closing oldest exchange ${exchangeToClose.id} for session ${sessionId}`);
+        exchangeToClose.close().catch(error => logger.error("Error closing exchange", error)); // TODO Promise??
     }
 }